+++ internal/webui/api_status_test.go
@@ -21,6 +21,7 @@ type statusFakeProvider struct {
 	update               health.UpdateInfo
 	watchers             map[string]bool
 	storeHealthy         *bool
+	attribution          *health.CorrelationAttribution
 }
 
 func (statusFakeProvider) Hostname() string { return "h" }
@@ -60,6 +61,9 @@ func (f statusFakeProvider) AutomationStatus() health.AutomationStatus {
 	return f.automation
 }
 func (f statusFakeProvider) UpdateInfo() health.UpdateInfo { return f.update }
+func (f statusFakeProvider) CorrelationAttribution() *health.CorrelationAttribution {
+	return f.attribution
+}
 
 var _ health.Provider = statusFakeProvider{}